+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 Change Log
 
+v2.5.0
+---
+* Improved hierarchy of generated directories when `--output` is a directory path
+* Fixed wrong obfuscated files path generation under `windows`. Fixed https://github.com/javascript-obfuscator/javascript-obfuscator/issues/576
+* Fixed wrong source map path generation under `windows`. Fixed https://github.com/javascript-obfuscator/javascript-obfuscator/issues/760
+* `javascript-obfuscator` now can be built under `windows` environment
+
 v2.4.3
 ---
 * Fixed https://github.com/javascript-obfuscator/javascript-obfuscator/issues/769